Make *Spec types taggable (#633)
* Proposal PR: Make *Spec types Taggable

This is a proposal PR to introduce tagging similar to OkHttp's tagging APIs. The idea here is that for an API that uses KotlinPoet as an IR, this allows consumers to tag arbitrary information to read at a later point. This also allows for pushing more information into individual specs (such as originating elements) without bolting them to the top-level API or coupling them to those frameworks.

One hypothetical example: Say Moshi's code gen was abstracted to be usable in non-apt environments. Right now there's a good deal of error handling and reporting in it that depends on having elements available to read from, but we could represent those elements as KotlinPoet types instead and track the original elements in tags for the processor's usage later.

From a high level - these are backed as a `Map<KClass<*>, Any>`. There's overloads in builders for `Class` and no-arg. There's no overload for the no-arg with `Any` though, as Java users will just see that as an `Object` anyway.

If this sounds reasonable, I can start on adding some tests.

* Remove parameter-less Any overloads

* Nix now-unnecessary JvmName annotations

* Taggable interface approach + keeping inline methods PoC

* Move builder to inside Taggable, clean docs, default taggable.tag(Class)

* Move buildTagMap() creator to Taggable

* Migrate rest to Taggable

* Constructor formatting

* Cast to nullable type

* Tag should be Any, not type var of the key type

* Differentiate key types and value types better

* Add TaggableTest

* Simplify parameters and use params as names

* Match types

* Fix style